requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
jinja2==3.1.2
google-cloud-firestore==2.13.1
google-cloud-storage==2.12.0
google-cloud-aiplatform==1.36.4
openai==1.3.5
anthropic==0.5.0
firebase-admin==6.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.6.1 